LayoutTests/platform/wpe/imported/w3c/web-platform-tests/encrypted-media/clearkey-mp4-playback-temporary-events-expected.txt


<<<<<<< HEAD
FAIL org.w3.clearkey, temporary, mp4, playback, check events assert_unreached: TypeError: Type error Reached unreachable code
=======
FAIL org.w3.clearkey, temporary, mp4, playback, check events assert_equals: Expected 1st event to be 'generaterequest' expected "generaterequest" but got "license-request"
>>>>>>> c80db0a1f694034dc600e650ee9956ed8d9279af
